requirements.txt
Flask==2.1.1
Pillow==9.2.0
pygame==2.1.2
libcamera